+++ ErsatzTV.Application/Resolutions/Commands/DeleteCustomResolution.cs
@@ -1,0 +1,5 @@
+using ErsatzTV.Core;
+
+namespace ErsatzTV.Application.Resolutions;
+
+public record DeleteCustomResolution(int ResolutionId) : IRequest<Option<BaseError>>;