+++ ErsatzTV.Core/Domain/Resolution.cs
@@ -8,6 +8,7 @@ public class Resolution : IDisplaySize
     public string Name { get; set; }
     public int Height { get; set; }
     public int Width { get; set; }
+    public bool IsCustom { get; set; }
 
     public override string ToString() => $"{Width}x{Height}";
 }